uservice-market/requirements.txt
grpcio-reflection==1.62.0
dapr-ext-grpc==1.12.0
onlexnet_market==0.0.10

# required as we have classes generated by avro-gen3
avro-gen3==0.7.12
fastavro==1.8.3

pytest==8.1.1

yfinance==0.2.36